[nvim] Add netrw file explorer configurations for nvim. Add some useful quick reference text files for regex and netrw usage.
Some of the options that are to be tried are 'set nocp' and 'filetype plugin on' might include it in the later commits if there are any problems. The listing style of netrw is not set to tree because there were some issues while copying and moving files to dirs.

Best practice so far is to open files in new tabs and use '<tab_no>gt' to switch tabs or just 'gt'. To mark multiple files 'V :norm mf'

The help texts only contain what is required for me now and it can be appended later.